+++ src/parse/mod.rs
@@ -5,8 +5,11 @@ mod value;
 
 use crate::expr::Expr;
 
-pub fn parse(input: &str) -> Expr {
-    let (remaining, expr) = expr::expr(input).unwrap();
-    assert!(remaining.is_empty());
-    expr
+impl Expr {
+    /// Parse an expression written in the Reval DSL format
+    pub fn parse(input: &str) -> Self {
+        let (remaining, expr) = expr::expr(input).unwrap();
+        assert!(remaining.is_empty());
+        expr
+    }
 }