+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Allow parsing individual `Expr` instances from json
 
+### Changed
+
+- Make `parse::parse` a method on `Expr`
+
 
 ## [0.2.0] - 2022-11-27